Add support for redirect with cookie authentication (#1)
* Add support for redirect with cookie authentication

In relation to feedreader/pluto#39, the request cookie gets saved for each redirect. This should support cookie based authentication.

* Support multiple cookies

Some websites return more than one cookie. This iterates over the multiple
cookies and concatenates them in a single one.

If only one cookie is returned, just pass that one to the next request.

* Make cookie header compliant with RFC 6265

RFC 6265 states that the client should pass a colon-separated cookie header
to the server as a single string. This commit should make the cookie request
compliant to this RFC.